+++ benchmark/general/xfs.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+
+if [[ "$kevin_env_set" != "1" ]]; then
+  echo "Please source env for kevin"
+  exit 1
+fi
 #usage: ./ext4_no_journal <device>
 
 sudo yes | mkfs.xfs -f $1
